is_irreducible
