+++ lie_nn/__init__.py
@@ -20,9 +20,8 @@ from ._src.clebsch_gordan import clebsch_gordan
 from ._src.infer_change_of_basis import infer_change_of_basis
 from ._src.conjugate import conjugate
 from ._src.real import make_explicitly_real, is_real
-from ._src.is_unitary import is_unitary
+from ._src.properties import is_unitary, is_irreducible
 from ._src.group_product import group_product
-from ._src.is_irreducible import is_irreducible
 
 from lie_nn import irreps as irreps
 from lie_nn import util as util
@@ -52,8 +51,8 @@ __all__ = [
     "make_explicitly_real",
     "is_real",
     "is_unitary",
-    "group_product",
     "is_irreducible",
+    "group_product",
     "irreps",
     "util",
     "finite",
